Move shouldForwardProp to lib/
